requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==2.5.1
requests==2.31.0
streamlit==1.25.0
scikit-learn==1.3.2
pandas==2.2.2
openai==1.4.0
pyDriller==0.10.0
lizard==1.23.3
shap==0.41.0
joblib==1.3.2

langchain==0.0.298
python-dotenv==1.0.1
redis==4.7.0
sqlalchemy==2.0.19
alembic==1.11.1
docker==6.1.3
boto3==1.34.0
PyJWT==2.8.0
hvac==1.1.1
